Made a lot of small changes here. Made two big changes though. One is that I changed it so that styles for each element are now stored in an associateive array where the selector for the style is the key into the array. This seemed a nice way to leverage Javascript's built in associative array functionality while opening the door for us to add styles based on classes or elements later on perhaps with some added logic. Essentially we've moved away from having styles inline. There are certainly a lot of disadvantages to this method as well and I'm open to discussing which one is better further.
Also, I changed all of our js files to use the module pattern. We seemed to have a lot of functions that we wanted to keep private and this seemed like one way of achieving that. Again, I'm not married to this solution and I don't mind changing the js files back to how they originally were if others don't like it too much.